pkp/pkp-lib#4158 Port categories code to pkp-lib
