+++ controllers/list/submissions/CatalogSubmissionsListHandler.inc.php
@@ -91,7 +91,7 @@ class CatalogSubmissionsListHandler extends SubmissionsListHandler {
 
 			$categories = array();
 			$categoryDao = DAORegistry::getDAO('CategoryDAO');
-			$categoriesResult = $categoryDao->getByPressId($context->getId());
+			$categoriesResult = $categoryDao->getByContextId($context->getId());
 			while (!$categoriesResult->eof()) {
 				$category = $categoriesResult->next();
 				list($categorySortBy, $categorySortDir) = explode('-', $category->getSortOption());
