+++ locale/de_DE/locale.xml
@@ -15,7 +15,6 @@
 	<!-- General monograph keys -->
 	<message key="monograph.audience">Zielgruppe</message>
 	<message key="monograph.coverImage">Titelbild</message>
-	<message key="monograph.currentCoverImage">Aktuelles Bild</message>
 	<message key="monograph.currentCoverImageReload">Speichern Sie Ihre Änderungen, um das neue Titelbild angezeigt zu bekommen.</message>
 	<message key="monograph.audience.rangeQualifier">Grundlage der Eingrenzung der Zielgruppe</message>
 	<message key="monograph.audience.rangeFrom">Bildungsstand (von)</message>
@@ -234,21 +233,6 @@
 	<message key="series.featured.description">Diese Reihen werden in der Hauptnavigation erscheinen</message>
 	<message key="series.path">Pfad</message>
 
-	<message key="grid.category.categories">Kategorien</message>
-	<message key="grid.category.add">Kategorie hinzufügen</message>
-	<message key="grid.category.edit">Kategorie bearbeiten</message>
-	<message key="grid.category.name">Name</message>
-	<message key="grid.category.path">Pfad</message>
-	<message key="grid.category.urlWillBe">Die URL der Kategorie soll lauten {$sampleUrl}</message>
-	<message key="grid.series.urlWillBe">Die URL der Reihe soll lauten {$sampleUrl}. 'Path' (deutsch "Pfad") ist eine Zeichenkette, durch die die Reihe eindeutig identifizierbar wird.</message>
-	<message key="grid.category.pathAlphaNumeric">Der Pfadname für eine Kategorie darf nur Buchstaben und Ziffern enthalten.</message>
-	<message key="grid.category.pathExists">Dieser Kategorienpfadname existiert bereits. Bitte geben Sie einen eindeutigen Pfadnamen ein.</message>
-	<message key="grid.category.description">Beschreibung</message>
-	<message key="grid.category.parentCategory">Übergeordnete Kategorie</message>
-	<message key="grid.category.removeText">Sind Sie sicher, dass Sie diese Kategorie entfernen möchten?</message>
-	<message key="grid.category.nameRequired">Bitte geben Sie einen Namen für die Kategorie ein.</message>
-	<message key="grid.category.categoryDetails">Genauere Angaben zur Kategorie</message>
-
 	<!-- Catalog Management -->
 	<message key="catalog.manage">Katalogverwaltung</message>
 	<message key="catalog.manage.homepage">Homepage</message>
@@ -303,10 +287,6 @@
 	<message key="catalog.sortBy.seriesDescription">Wählen Sie aus, in welcher Reihenfolge Bücher in dieser Reihe angezeigt werden sollen.</message>
 	<message key="catalog.sortBy.categoryDescription">Wählen Sie aus, in welcher Reihenfolge Bücher in dieser Kategorie angezeigt werden sollen.</message>
 	<message key="catalog.sortBy.catalogDescription">Wählen Sie aus, in welcher Reihenfolge Bücher im Katalog angezeigt werden sollen.</message>
-	<message key="catalog.sortBy.titleAsc">Titel (A-Z)</message>
-	<message key="catalog.sortBy.titleDesc">Titel (Z-A)</message>
-	<message key="catalog.sortBy.datePublishedAsc">Erscheinungsdatum (älteste zuerst)</message>
-	<message key="catalog.sortBy.datePublishedDesc">Erscheinungsdatum (neueste zuerst)</message>
 	<message key="catalog.sortBy.seriesPositionAsc">Position in Reihe (niedrigste zuerst)</message>
 	<message key="catalog.sortBy.seriesPositionDesc">Position in Reihe (höchste zuerst)</message>
 	<message key="catalog.viewableFile.title">{$type}-Ansicht der Datei {$title}</message>
@@ -577,4 +557,5 @@ Ihr Server untersützt aktuell mbstring: <strong>{$supportsMBString}</strong>]]>
 
 	<message key="user.profile.form.showOtherContexts">Andere Verlage anzeigen</message>
 	<message key="user.profile.form.hideOtherContexts">Andere Verlage verbergen</message>
+	<message key="grid.series.urlWillBe">Die URL der Reihe soll lauten {$sampleUrl}. 'Path' (deutsch "Pfad") ist eine Zeichenkette, durch die die Reihe eindeutig identifizierbar wird.</message>
 </locale>
