+++ locale/en_US/locale.xml
@@ -15,7 +15,6 @@
 	<!-- General monograph keys -->
 	<message key="monograph.audience">Audience</message>
 	<message key="monograph.coverImage">Cover Image</message>
-	<message key="monograph.currentCoverImage">Current Image</message>
 	<message key="monograph.currentCoverImageReload">Save your changes to view the new cover image.</message>
 	<message key="monograph.audience.rangeQualifier">Audience Range Qualifier</message>
 	<message key="monograph.audience.rangeFrom">Audience Range (from)</message>
@@ -294,10 +293,6 @@
 	<message key="catalog.sortBy.seriesDescription">Choose how to order books in this series.</message>
 	<message key="catalog.sortBy.categoryDescription">Choose how to order books in this category.</message>
 	<message key="catalog.sortBy.catalogDescription">Choose how to order books in the catalog.</message>
-	<message key="catalog.sortBy.titleAsc">Title (A-Z)</message>
-	<message key="catalog.sortBy.titleDesc">Title (Z-A)</message>
-	<message key="catalog.sortBy.datePublishedAsc">Publication date (oldest first)</message>
-	<message key="catalog.sortBy.datePublishedDesc">Publication date (newest first)</message>
 	<message key="catalog.sortBy.seriesPositionAsc">Series position (lowest first)</message>
 	<message key="catalog.sortBy.seriesPositionDesc">Series position (highest first)</message>
 	<message key="catalog.viewableFile.title">{$type} view of the file {$title}</message>
@@ -582,5 +577,4 @@ Your server currently supports mbstring: <strong>{$supportsMBString}</strong>]]>
 
 	<message key="user.profile.form.showOtherContexts">Register with other presses</message>
 	<message key="user.profile.form.hideOtherContexts">Hide other presses</message>
-	<message key="grid.category.categories">Categories</message>
 </locale>
