+++ plugins/blocks/browse/BrowseBlockPlugin.inc.php
@@ -108,7 +108,7 @@ class BrowseBlockPlugin extends BlockPlugin {
 		if ($categoriesDisplay) {
 			// Provide a list of categories to browse
 			$categoryDao = DAORegistry::getDAO('CategoryDAO');
-			$categories = $categoryDao->getByPressId($press->getId());
+			$categories = $categoryDao->getByContextId($press->getId());
 			$templateMgr->assign('browseCategoryFactory', $categories);
 		}
